backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai
langchain-core
python-dotenv
pillow